Refresh the list of servers to monitor at runtime
An important bit of this commit is that the particular instances
that we are using in EvmServer are preserved by the refresh process.
We can't just treat it as a cache and overwrite the local list.

This is because a bunch of each server's state is stored in instance
variables. If we were to overwrite the instance we would lose all
of that information which would cause (unquestionably bad) undefined
behavior.