Don't thrash settings when we only have one server
Before this change, a single server would init the settings every
time it was monitored. This causes an unnecessary amount of extra
work in the appliance model where we will always have only a single
server.

To avoid this, return from #impersonate_server if we already are
the requested server, and only reset the caches if we changed servers.