Fix specs to handle extra server creation when podified
In specs we often will create a server, which creates a zone for
us in the factory. Previously, even when we were simulating running
in pods, that would leave us with one server and one zone.

Now, the zone creation process is checking the environment for
podified and creating a server itself. This leaves us with an
extra call to MiqEnvironment::Command.is_podified and some bonus
server records.

This commit handles both of those issues.