Format the label for the zone correctly
Without this change we get errors from Kube about the label format:

$> oc label node node.example.com manageiq/zone/asdf=true
The Node "node.example.com" is invalid: metadata.labels: Invalid value:
"manageiq/zone/asdf": a qualified name must consist of alphanumeric
characters, '-', '_' or '.', and must start and end with an
alphanumeric character (e.g. 'MyName',  or 'my.name',  or '123-abc',
regex used for validation is '([A-Za-z0-9][-A-Za-z0-9_.]*)?[A-Za-z0-9]')
with an optional DNS subdomain prefix and '/' (e.g. 'example.com/MyName')