Don't start the heartbeat worker in do_before_work_loop
The heartbeat worker thread and the main thread were competing
to define the Workers::MiqDefaults constant.

The main thread calls heartbeat from do_work_loop which is
unavoidable. So to solve the problem we just wait until the first
call to do_work to start the heartbeat worker.

This allows the main thread to resolve the constant without contention
and the heartbeat worker still gets started before we process the
first queue message

Fixes #19813